dsds

+++ ft_atoi.c
@@ -6,15 +6,17 @@
 /*   By: zbakour <zbakour@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/24 17:57:34 by zbakour           #+#    #+#             */
-/*   Updated: 2024/10/24 18:20:54 by zbakour          ###   ########.fr       */
+/*   Updated: 2024/11/10 21:20:01 by zbakour          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 int	ft_atoi(const char *nptr)
 {
-	int	sign;
-	int	i;
-	int	total;
+	int		sign;
+	int		i;
+	long	total;
 
 	total = 0;
 	i = 0;
@@ -34,5 +36,5 @@ int	ft_atoi(const char *nptr)
 		total = total * 10 + (nptr[i] - '0');
 		i++;
 	}
-	return (total * sign);
+	return ((int)total * sign);
 }
